RDP: Refactor protocol code
