+++ lib/msf/core/exploit/mixins.rb
@@ -70,6 +70,7 @@ require 'msf/core/exploit/realport'
 require 'msf/core/exploit/sip'
 require 'msf/core/exploit/tincd'
 require 'msf/core/exploit/git'
+require 'msf/core/exploit/rdp'
 
 # Telephony
 require 'msf/core/exploit/dialup'
